zsh-completions: correct instructions for fixing compaudit
Issue #121742 details an issue with the instructions at the end of the zsh-completions formula. Specifically, there is an issue with directory permissions that the instructions suggest fixing, however those instructions are incomplete. The permissions must be fixed recursively for `/opt/homebrew/share/zsh` _as well as_ for `/opt/homebrew/share` (non-recursively).